+++ src/app/components/location/forms/location-form-light/location-form-light.component.spec.ts
@@ -1,36 +1,14 @@
-import { EventEmitter } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TranslocoService } from '@ngneat/transloco';
-import { LocationService } from '@services/location.service';
-import { MockPipe, MockProvider, MockProviders } from 'ng-mocks';
-import { of } from 'rxjs';
-
+import { MockBuilder } from 'ng-mocks';
 import { LocationFormLightComponent } from './location-form-light.component';
 
 describe('LocationFormLightComponent', () => {
-  let component: LocationFormLightComponent;
-  let fixture: ComponentFixture<LocationFormLightComponent>;
 
   beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        LocationFormLightComponent,
-        // MockPipe(TranslocoService, value => value)
-      ],
-      providers: [
-        ...MockProviders(LocationService,TranslocoService),
-      ],
-      declarations: [
-      ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(LocationFormLightComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    return MockBuilder(LocationFormLightComponent)
   });
 
   it('should create', () => {
+    const component = LocationFormLightComponent;
     expect(component).toBeTruthy();
   });
 });