+++ src/app/components/location/forms/location-form-name/location-form-name.component.spec.ts
@@ -1,29 +1,13 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { TranslocoModule } from '@ngneat/transloco';
-
+import { MockBuilder } from 'ng-mocks';
 import { LocationFormNameComponent } from './location-form-name.component';
 
 describe('LocationFormNameComponent', () => {
-  let component: LocationFormNameComponent;
-  let fixture: ComponentFixture<LocationFormNameComponent>;
-
   beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [
-        LocationFormNameComponent,
-        TranslocoModule,
-        BrowserAnimationsModule
-      ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(LocationFormNameComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    return MockBuilder(LocationFormNameComponent);
   });
 
   it('should create', () => {
+    const component = LocationFormNameComponent;
     expect(component).toBeTruthy();
   });
 });