+++ src/app/components/navigation/theme-switcher/theme-switcher.component.spec.ts
@@ -1,23 +1,14 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { MockBuilder } from 'ng-mocks';
+import { TranslocoService } from '@ngneat/transloco';
 import { ThemeSwitcherComponent } from './theme-switcher.component';
 
 describe('ThemeSwitcherComponent', () => {
-  let component: ThemeSwitcherComponent;
-  let fixture: ComponentFixture<ThemeSwitcherComponent>;
-
   beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [ThemeSwitcherComponent]
-    })
-    .compileComponents();
-    
-    fixture = TestBed.createComponent(ThemeSwitcherComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    return MockBuilder(ThemeSwitcherComponent).mock(TranslocoService);
   });
 
   it('should create', () => {
+    const component = ThemeSwitcherComponent;
     expect(component).toBeTruthy();
   });
 });