+++ src/app/components/plant/menu/plant-menu-water/plant-menu-water.component.spec.ts
@@ -1,21 +1,14 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { PlantMenuWaterComponent } from './plant-menu-water.component';
+import { MockBuilder } from 'ng-mocks';
 
 describe('PlantMenuWaterComponent', () => {
-  let component: PlantMenuWaterComponent;
-  let fixture: ComponentFixture<PlantMenuWaterComponent>;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [PlantMenuWaterComponent]
-    });
-    fixture = TestBed.createComponent(PlantMenuWaterComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    return MockBuilder(PlantMenuWaterComponent);
   });
 
   it('should create', () => {
+    const component = PlantMenuWaterComponent;
     expect(component).toBeTruthy();
   });
 });