+++ src/app/components/plant/plant-base-action/plant-base-action.component.spec.ts
@@ -1,21 +1,15 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { PlantBaseActionComponent } from './plant-base-action.component';
+import { MockBuilder } from 'ng-mocks';
+import { PlantService } from '@services/plant.service';
 
 describe('PlantBaseActionComponent', () => {
-  let component: PlantBaseActionComponent;
-  let fixture: ComponentFixture<PlantBaseActionComponent>;
-
   beforeEach(() => {
-    TestBed.configureTestingModule({
-      imports: [PlantBaseActionComponent]
-    });
-    fixture = TestBed.createComponent(PlantBaseActionComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
+    return MockBuilder(PlantBaseActionComponent)
+      .mock(PlantService);
   });
 
   it('should create', () => {
+    const component = PlantBaseActionComponent;
     expect(component).toBeTruthy();
   });
 });