+++ src/app/services/share.service.spec.ts
@@ -1,0 +1,15 @@
+import { MockBuilder } from 'ng-mocks';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { ShareService } from './share.service';
+
+describe('ShareService', () => {
+  beforeEach(() => {
+    return MockBuilder(ShareService)
+      .keep(HttpClientTestingModule)
+  });
+
+  it('should be created', () => {
+    const service = ShareService;
+    expect(service).toBeTruthy();
+  });
+});