+++ src/app/tests/transloco.module.ts
@@ -1,0 +1,18 @@
+import {
+  TranslocoTestingModule,
+  TranslocoTestingOptions,
+} from '@ngneat/transloco';
+import en from 'src/assets/i18n/en.json';
+import es from 'src/assets/i18n/es.json';
+
+export function getTranslocoModule(options: TranslocoTestingOptions = {}) {
+  return TranslocoTestingModule.forRoot({
+    langs: { en, es },
+    translocoConfig: {
+      availableLangs: ['en', 'es'],
+      defaultLang: 'en',
+    },
+    preloadLangs: true,
+    ...options,
+  });
+}